requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-cors-headers==3.2.0
django-filter==2.2.0
django-guardian==2.1.0
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.2.1
pycryptodome==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
